@@ -1,48 +1,33 @@
 {
   "name": "control-records",
-  "version": "0.1.0",
   "private": true,
+  "version": "1.0.0",
   "scripts": {
-    "start": "npm run serve",
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "start": "vite",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint --ext .js,.vue src"
   },
   "dependencies": {
-    "@kangc/v-md-editor": "1.6.4",
-    "codemirror": "5.61.0",
-    "core-js": "3.12.1",
-    "dompurify": "2.2.8",
-    "jspdf": "2.3.1",
-    "marked": "2.0.3",
-    "vue": "2.6.12"
+    "@kangc/v-md-editor": "^2.3.15",
+    "codemirror": "^5.65.6",
+    "dompurify": "^2.3.8",
+    "jspdf": "^2.5.1",
+    "marked": "^4.0.17",
+    "vue": "^3.2.25"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "4.5.13",
-    "@vue/cli-plugin-eslint": "4.5.13",
-    "@vue/cli-service": "4.5.13",
-    "babel-eslint": "10.1.0",
-    "eslint": "6.8.0",
-    "eslint-plugin-vue": "7.9.0",
-    "vue-template-compiler": "2.6.12"
+    "@vitejs/plugin-vue": "^2.3.3",
+    "eslint": "^8.19.0",
+    "eslint-plugin-vue": "^9.2.0",
+    "vite": "^2.9.9"
   },
   "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
     "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended"
+      "eslint:recommended",
+      "plugin:vue/vue3-essential"
     ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
     "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+  }
 }
