@@ -1,11 +1,11 @@
-import Vue from 'vue'
-import './click-outside.directive.js'
+import { createApp } from 'vue'
 import App from './App.vue'
 
 // Import Vue Markdown Editor
-import VueMarkdownEditor from '@kangc/v-md-editor/lib/codemirror-editor'
+import VMdEditor from '@kangc/v-md-editor/lib/codemirror-editor'
 import '@kangc/v-md-editor/lib/style/codemirror-editor.css'
-import vuepressTheme from '@kangc/v-md-editor/lib/theme/vuepress.js'
+import githubTheme from '@kangc/v-md-editor/lib/theme/github.js'
+import '@kangc/v-md-editor/lib/theme/style/github.css'
 import enUS from '@kangc/v-md-editor/lib/lang/en-US'
 
 // Import CodeMirror
@@ -17,12 +17,9 @@ import 'codemirror/addon/scroll/simplescrollbars'
 import 'codemirror/addon/scroll/simplescrollbars.css'
 import 'codemirror/lib/codemirror.css'
 
-// Configure Vue
-Vue.config.productionTip = false
-
 // Setup Vue Markdown Editor
-VueMarkdownEditor.lang.use('en-US', enUS)
-VueMarkdownEditor.lang.add({
+VMdEditor.lang.use('en-US', enUS)
+VMdEditor.lang.add({
   'en-US': {
     italic: {
       toolbar: 'Underline',
@@ -30,11 +27,10 @@ VueMarkdownEditor.lang.add({
     },
   }
 })
-VueMarkdownEditor.Codemirror = Codemirror
-VueMarkdownEditor.use(vuepressTheme)
-Vue.use(VueMarkdownEditor)
+VMdEditor.Codemirror = Codemirror
+VMdEditor.use(githubTheme)
 
-// Initialize Vue
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App)
+
+app.use(VMdEditor)
+app.mount('#app')
